made ECM activatable

+++ src/cmd/images.h
@@ -60,7 +60,7 @@ struct UnitImages {
 
   float *cockpit_damage; //0 is radar, 1 to MAXVDU is vdus and >MAXVDU is gauges
   ///how likely to fool missiles
-  float ecm;
+  short ecm;
   ///holds the info for the repair bot type. 0 is no bot;
   unsigned char repair_droid;
   ///How much energy cloaking takes per frame
